+++ src/library/DIPS.Mobile.UI/API/Builder/AppHostBuilderExtensions.cs
@@ -77,6 +77,7 @@ public static partial class AppHostBuilderExtensions
             effects.Add<FadeInEffect, FadeInPlatformEffect>();
             effects.Add<FadeOutEffect, FadeOutPlatformEffect>();
             effects.Add<ScaleInEffect, ScaleInPlatformEffect>();
+            effects.Add<Effects.Accessibility.Effects.TraitEffect, Effects.Accessibility.Effects.TraitPlatformEffect>();
         });
 
         builder.UseSkiaSharp();