+++ src/library/DIPS.Mobile.UI/Resources/LocalizedStrings/LocalizedStrings/DUILocalizedStrings.Designer.cs
@@ -452,5 +452,77 @@ namespace DIPS.Mobile.UI.Resources.LocalizedStrings.LocalizedStrings {
                 return ResourceManager.GetString("Accessability_TapToActivateFlash", resourceCulture);
             }
         }
+        
+        internal static string Selected {
+            get {
+                return ResourceManager.GetString("Selected", resourceCulture);
+            }
+        }
+        
+        internal static string NotSelected {
+            get {
+                return ResourceManager.GetString("NotSelected", resourceCulture);
+            }
+        }
+        
+        internal static string Accessability_SelectSortOption {
+            get {
+                return ResourceManager.GetString("Accessability_SelectSortOption", resourceCulture);
+            }
+        }
+        
+        internal static string Ascending {
+            get {
+                return ResourceManager.GetString("Ascending", resourceCulture);
+            }
+        }
+        
+        internal static string Descending {
+            get {
+                return ResourceManager.GetString("Descending", resourceCulture);
+            }
+        }
+        
+        internal static string Accessability_SelectedSort {
+            get {
+                return ResourceManager.GetString("Accessability_SelectedSort", resourceCulture);
+            }
+        }
+        
+        internal static string Accessability_TapToChangeSorting {
+            get {
+                return ResourceManager.GetString("Accessability_TapToChangeSorting", resourceCulture);
+            }
+        }
+        
+        internal static string InputField {
+            get {
+                return ResourceManager.GetString("InputField", resourceCulture);
+            }
+        }
+        
+        internal static string Dictation {
+            get {
+                return ResourceManager.GetString("Dictation", resourceCulture);
+            }
+        }
+        
+        internal static string CounterValue {
+            get {
+                return ResourceManager.GetString("CounterValue", resourceCulture);
+            }
+        }
+        
+        internal static string Urgent {
+            get {
+                return ResourceManager.GetString("Urgent", resourceCulture);
+            }
+        }
+        
+        internal static string Header {
+            get {
+                return ResourceManager.GetString("Header", resourceCulture);
+            }
+        }
     }
 }